my_openai/AgentDemoUseMCP/mcp-agent-project/requirements.txt
Flask==2.0.1
requests==2.25.1
pytest==6.2.4
python-dotenv==0.19.1
google-generativeai==0.1.0
openai==0.27.0